Move remaining shared utilities into utils directories
