+++ kernel/include/ARPContext.h
@@ -29,7 +29,7 @@
 #include "Device.h"
 #include "Network.h"
 #include "Endianness.h"
-#include "Notification.h"
+#include "utils/Notification.h"
 #include "utils/AdaptiveDelay.h"
 
 /************************************************************************/
