+++ kernel/include/Kernel.h
@@ -38,7 +38,7 @@
 #include "Multiboot.h"
 #include "Process.h"
 #include "String.h"
-#include "TOML.h"
+#include "utils/TOML.h"
 #include "Cache.h"
 #include "Text.h"
 #include "User.h"
