+++ kernel/include/utils/Hysteresis.h
