+++ kernel/source/SYSCall.c
@@ -27,7 +27,7 @@
 #include "../include/Console.h"
 #include "../include/File.h"
 #include "../include/Heap.h"
-#include "../include/Helpers.h"
+#include "../include/utils/Helpers.h"
 #include "../include/ID.h"
 #include "../include/Kernel.h"
 #include "../include/Keyboard.h"
