+++ kernel/source/Shell.c
@@ -33,14 +33,14 @@
 #include "../include/FileSystem.h"
 #include "../include/GFX.h"
 #include "../include/Heap.h"
-#include "../include/Helpers.h"
+#include "../include/utils/Helpers.h"
 #include "../include/Kernel.h"
 #include "../include/Keyboard.h"
 #include "../include/List.h"
 #include "../include/Log.h"
 #include "../include/Network.h"
 #include "../include/NetworkManager.h"
-#include "../include/Path.h"
+#include "../include/utils/Path.h"
 #include "../include/Process.h"
 #include "../include/Script.h"
 #include "../include/StackTrace.h"
