+++ kernel/source/UserAccount.c
@@ -28,7 +28,7 @@
 #include "../include/utils/Crypt.h"
 #include "../include/utils/Database.h"
 #include "../include/Heap.h"
-#include "../include/Helpers.h"
+#include "../include/utils/Helpers.h"
 #include "../include/Kernel.h"
 #include "../include/List.h"
 #include "../include/Log.h"
