+++ kernel/source/UserSession.c
@@ -26,7 +26,7 @@
 
 #include "../include/Clock.h"
 #include "../include/Heap.h"
-#include "../include/Helpers.h"
+#include "../include/utils/Helpers.h"
 #include "../include/Kernel.h"
 #include "../include/List.h"
 #include "../include/Log.h"
