+++ kernel/source/utils/TOML.c
@@ -22,7 +22,7 @@
 
 \************************************************************************/
 
-#include "../include/TOML.h"
+#include "../include/utils/TOML.h"
 
 #include "../include/Kernel.h"
 #include "../include/Log.h"
