+++ kernel/source/utils/UUID.c
@@ -21,7 +21,7 @@
 
 \************************************************************************/
 
-#include "../include/UUID.h"
+#include "../include/utils/UUID.h"
 
 /***************************************************************************/
